backend/requirements.txt
click==8.1.3
colorama==0.4.5
deta==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
